28-11-2016

+++ src/main/java/com/example/repository/UserRepository.java
@@ -1,10 +1,11 @@
 package com.example.repository;
 
 import com.example.entity.Users;
+import com.sun.org.apache.xpath.internal.operations.String;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<Users, Long> {
+public interface UserRepository extends JpaRepository<Users, Integer> {
 
 }
